Apple doesn't allow to open a kernel socket during pr_init(). For now: * disable SCTP_APPLE_AUTO_ASCONF * Use 0 as the default local encapsulation port
When things are stable, we can use a small timer
to get the above running again.

This change allows the NKE to be loaded and unloaded.
However, sending packets does not work and killing
an endpoint will crash the system. Apple has changed
the freeing of the inps/pcbs.

Getting packets in and out is the next on the ToDo list.